Refactor: Updated folder structure + remove unused code
