+++ StringCalculator1/Services/StringParser/IStringParser.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace StringCalculator1.Services
+namespace StringCalculator1.Services.StringParser
 {
     public interface IStringParser
     {
