+++ StringCalculator1Tests/Services/DelimeterServiceTests.cs
@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using StringCalculator1;
-using StringCalculator1.Services;
-using StringCalculator1.Services.Delimeter;
+﻿using StringCalculator1.Services.DelimeterService;
 
 namespace StringCalculator1Tests.Services
 {
