+++ StringCalculator1Tests/StringCalculator1Tests.cs
@@ -1,6 +1,6 @@
 ﻿using Moq;
 using StringCalculator1;
-using StringCalculator1.Services;
+using StringCalculator1.Services.StringParser;
 
 namespace StringCalculator1Tests
 {
